% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareEstPar.R
\name{compareEstPar}
\alias{compareEstPar}
\title{compareEstPar}
\usage{
compareEstPar(pars)
}
\arguments{
\item{estimatePars}{an estimatePar object}
}
\description{
Compare estimated parameters with the simulated "truth"
}
