% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOccNorm.R
\name{createOccNorm}
\alias{createOccNorm}
\title{createOccNorm}
\usage{
createOccNorm(map, intercept = 0.5, sigmaFylke = 0.1, sigmaKommune = 0.1,
  sigmaGrid = 0, sigmaFylkeTrend = 0, sigmaKommuneTrend = 0,
  artypeEff = c(Bebygd = 0, Samferdsel = 0, `Fulldyrka jord` = 0,
  `Overflatedyrka jord` = 0, Innmarksbeite = 0, Skog = 0, `Åpen fastmark` = 0,
  Myr = 0, Isbre = 0, Ferskvann = 0, Hav = 0, `Ikke kartlagt` = 0),
  artypeTrend = c(Bebygd = 0, Samferdsel = 0, `Fulldyrka jord` = 0,
  `Overflatedyrka jord` = 0, Innmarksbeite = 0, Skog = 0, `Åpen fastmark` = 0,
  Myr = 0, Isbre = 0, Ferskvann = 0, Hav = 0, `Ikke kartlagt` = 0),
  nYears = 5, interceptTrend = -0.05, sdInterceptTrend = 0,
  sortFylke = T, sortKommune = T, sortGrid = T)
}
\arguments{
\item{sigmaFylke}{Standard deviance of fylkes}

\item{formula}{Not yet implemented}

\item{sigmaFylke}{Standard deviance of kommunes.}

\item{sigmaFylke}{Standard deviance of grids.}
}
\description{
Creates a probability of species occurrence based on explanatory factors.
The levels are calculated additively with a fylke effect, a kommune effect, plus a random part (sigma) that
determines the intra-kommune variation among grid cells.
This is intended to fill a "map" with values of occurrences, and expects columns named fylke and kommune.
}
\examples{
\dontrun{

}
}
