% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOccPois.R
\name{createOccPois}
\alias{createOccPois}
\title{createOccPois}
\usage{
createOccPois(map, intercept = 22, formula = ~NULL, sigmaFylke = 0.1,
  sigmaKommune = 0.1, sigmaGrid = 0, sigmaFylkeTrend = 0,
  sigmaKommuneTrend = 0, artypeEff = c(Bebygd = 1, Samferdsel = 1,
  `Fulldyrka jord` = 1, `Overflatedyrka jord` = 1, Innmarksbeite = 1, Skog = 1,
  `Åpen fastmark` = 1, Myr = 1, Isbre = 1, Ferskvann = 1, Hav = 1, `Ikke
  kartlagt` = 1), artypeTrend = c(Bebygd = 1, Samferdsel = 1, `Fulldyrka jord`
  = 1, `Overflatedyrka jord` = 1, Innmarksbeite = 1, Skog = 1, `Åpen fastmark`
  = 1, Myr = 1, Isbre = 1, Ferskvann = 1, Hav = 1, `Ikke kartlagt` = 1),
  nYears = 5, interceptTrend = 0.95, sdInterceptTrend = 0,
  sortFylke = T, sortKommune = T, sortGrid = T)
}
\arguments{
\item{formula}{Not yet implemented}

\item{sigmaFylke}{Standard deviance of fylkes}

\item{artypeEff}{Note that all parameters are input on the original scale (easier than using log scale)}

\item{sigmaFylke}{Standard deviance of kommunes.}

\item{sigmaFylke}{Standard deviance of grids.}
}
\description{
Creates a lambda of species counts (for use in a Poisson draw) based on explanatory factors.
The levels are calculated additively with a fylke effect, a kommune effect, plus a random part (sigma) that
determines the intra-kommune variation among grid cells.
This is intended to fill a "map" with values of occurrences, and expects columns named fylke and kommune.
}
\examples{
\dontrun{

}
}
