% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOccProb.R
\name{createOccProb}
\alias{createOccProb}
\title{createOccProb}
\usage{
createOccProb(map, intercept = 0.1, formula = ~lat + lon,
  fylkeParams = c(0.1, 0.4), kommuneParams = c(0.1, 0.4), sigma = 0,
  scale = T)
}
\arguments{
\item{formula}{}
}
\description{
Creates a probability of species occurrence based on explanatory factors.
The levels are calculated additively with a fylke effect, a kommune effect, plus a random part (sigma) that
determines the intra-kommune variation among grid cells.
This is intended to fill a "map" with values of occurrences, and expects columns named fylke and kommune.
}
\examples{
\dontrun{

}
}
