% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectProb.R
\name{detectProp}
\alias{detectProp}
\title{detectProb}
\usage{
detectProp(occProb, detectProb, locations, visits, visitCost, nBoot)
}
\arguments{
\item{occProb}{vector of occurence probabilities to test}

\item{detectProb}{vector of detection probabilities}

\item{locations}{vector of number of locations to visit/sample}

\item{visits}{vector of number of visits per sample}

\item{visitCost}{vector of costs of each visit}

\item{nBoot}{number of bootstrap samples to calculate std error of percent detected}
}
\description{
Function that produces empirical estimates of probability to detect a foreign species, and associated costs
based on varying occurrences, detectabilities and survey schemes.
}
\examples{
{
tt <- detectProp(occProb = seq(0.01, 0.9, by = 0.05),
detectProb = 0.8,
locations = 30,
visits =  4,
visitCost = 10,
nBoot =  9999)

}

}
