% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePars.R
\name{estimatePars}
\alias{estimatePars}
\title{estimatePars}
\usage{
estimatePars(map = NULL, sampleFun = c("sampleNorm", "sampleBin",
  "samplePois"), samplePars = list(yearlyCapacity = 100),
  model = c("normalFull", "binomialFull", "poissonFull"), nIter = 1)
}
\arguments{
\item{sampleFun}{which sampling function to use}

\item{samplePars}{an optional list of parameters for the sampling function, such as resampling times etc}

\item{model}{which model to fit. Currently only full models of three types is supported. Could perhaps be more flexible.}

\item{population}{a surveyHat object, containing the realised mean occurrences and population parameters}
}
\description{
Fits a model to the drawn survey data that estimates the population parameters and compares these to the true population parameters.
}
\examples{
\dontrun{

}
}
