% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expEnd.R
\name{expEnd}
\alias{expEnd}
\title{expEnd}
\usage{
expEnd(N0 = 100, lambda, T, perc = T)
}
\arguments{
\item{N0}{Initial amount}

\item{lambda}{yearly rate of change}

\item{T}{timespan to calculate over}

\item{perc}{should the result be presented as a percentage of the initial value? True/False}
}
\description{
Calculates the resulting number from a initial number given a yearly decay rate, given an exponential change function
}
