% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSSBGrid.R
\name{getSSBGrid}
\alias{getSSBGrid}
\title{getSSBGrid}
\usage{
getSSBGrid(conn = con, scale = c("10km", "5km", "1km", "500m"))
}
\arguments{
\item{scale}{Scale of SSB grid to retrieve. Options are 10km, 5km, 1km and 500m.
Note that smaller scales are time consuming!}
}
\description{
Retrieves the SSB grids, together with fylke and kommune names (and ids) overlapping the center of each grid cell, as an sf object.
}
\examples{
\dontrun{
fylke10km <- getSSBGrid()
plot(fylke10km["fylke"])
}
}
