% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map10km}
\alias{map10km}
\title{map10km}
\format{Simple feature collection with 4057 features and 8 fields}
\source{
\code{map10km <- getSSBGrid(conn = con, scale = "10km")}
}
\usage{
map10km
}
\description{
Map of SSB-grids with AR5 artype on a 10km scale grid.
This could as easily be gotten from the function below, but is stored here for speed.
}
\keyword{datasets}
