% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.surveyAlt.R
\name{plot.surveyAlt}
\alias{plot.surveyAlt}
\title{plotSurveyAlt}
\usage{
\method{plot}{surveyAlt}(df, grouping = "yearlyCapacity",
  color = c("yearlyCapacity", "totSamples", "resampleTime"),
  allTicks = T)
}
\description{
Plot function for
}
