% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresConnect.R
\name{postgresConnect}
\alias{postgresConnect}
\title{postgresConnect}
\usage{
postgresConnect(Username = "postgjest", Password = "gjestpost",
  host = "ninpgsql03.nina.no")
}
\arguments{
\item{host}{Name of database server. Defaults to (new) main PostgreSQL machine "ninpgsql03.nina.no".}

\item{user}{Username, defaults to guest account with basic read privileges}

\item{password}{Password Beware of storing this in cleartext in scripts. Consider storing it in a file that is included in gitignore.}
}
\description{
postgresConnect
}
