% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAlternatives.R
\name{sampleAlternatives}
\alias{sampleAlternatives}
\title{sampleAlternatives}
\usage{
sampleAlternatives(maxTime, maxCapacity, stepsCapacity = NULL)
}
\arguments{
\item{maxTime}{Total number of years of the study. Reasonably between 5 - 100 years.}

\item{maxCapacity}{Maximum number of localities possible to survey each year.}

\item{stepsCapacity}{Interval step size of number of localities per year. E.g. calculate possible regimes with 10, 20, and 30 number of localities per year. Defaults to maxCapacity.}
}
\value{
A tibble with potential survey regimes.
}
\description{
Function to calculate possible regimes in terms of number of replicates, number of localities and number of replicates per locality.
This calculates possible survey regimes given the possible resampling times that results in a balanced data set. This hinges on only
two parameters, the total timespan of the entire survey, the yearly survey capacity of locations.
}
