% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBin.R
\name{sampleBin}
\alias{sampleBin}
\title{sampleBin}
\usage{
sampleBin(map, column, subFylke = NULL, subKommune = NULL, rows = NULL,
  nSites = NULL, nObs = 1)
}
\arguments{
\item{column}{Column to draw samples from.}

\item{rows}{Vector of which map grid cells (row numbers) to draw from, for repeated samples from the same grid cells. Incompatible with n.}

\item{nSites}{Integer, number of new random grids to sample. Imcompatible with rows.}

\item{nObs}{Number of observations to draw. Can be a single integer or a vector of integers. This will be coerced to integer(s).}
}
\description{
sampleBin
}
\examples{
\dontrun{

}
}
