% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNorm.R
\name{sampleNorm}
\alias{sampleNorm}
\title{sampleNorm}
\usage{
sampleNorm(map, column, subFylke = NULL, subKommune = NULL,
  gridCells = NULL, nSites = NULL, sampleErr = 0)
}
\arguments{
\item{column}{Column to draw samples from.}

\item{gridCells}{Vector of which map grid cells to draw from, for repeated samples from the same grid cells. Incompatible with n.}

\item{nSites}{Integer, number of new random grids to sample. Imcompatible with rows.}

\item{sampleErr}{Double precision. Sampling error as standard deviation of a normally distributed variable.}
}
\description{
sampleNorm
}
\examples{
\dontrun{

}
}
