% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePois.R
\name{samplePois}
\alias{samplePois}
\title{samplePois}
\usage{
samplePois(map, column, subFylke = NULL, subKommune = NULL,
  gridCells = NULL, nSites = NULL)
}
\arguments{
\item{column}{Column to draw samples from.}

\item{gridCells}{Vector of which map grid cells to draw from, for repeated samples from the same grid cells. Incompatible with n.}

\item{nSites}{Integer, number of new random grids to sample. Imcompatible with rows.}

\item{sampleErr}{Double precision. Sampling error as standard deviation of a normally distributed variable.}
}
\description{
samplePois
}
\examples{
\dontrun{

}
}
