% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedDetection.R
\name{weightedDetection}
\alias{weightedDetection}
\title{weightedDetection}
\usage{
weightedDetection(occWeights = data.frame(sites = 1:10, weights = c(0.05,
  0.05, 0.05, 0.05, 0.1, 0.1, 0.1, 0.1, 0.2, 0.2)),
  visWeights = data.frame(sites = 1:10, weights = c(0.05, 0.05, 0.05, 0.05,
  0.1, 0.1, 0.1, 0.1, 0.2, 0.2)), noOccur = 5, noLocations = 5,
  noVisits = 5, detectProb = 0.1, nIter = 999)
}
\arguments{
\item{occWeights}{occurrenceWeights}
}
\description{
Calculate the probability of detecting a species with weighted occurrences and visits
}
\examples{
\dontrun{

tt <- weightedDetection(noOccur = 1,
noLocations = 1:10,
detectProb =  0.5)

plot(tt)

occ <- createOccProb(map10km)

visWeights <- occ \%>\%
 select(sites = ssbid,
         weights = prob)

system.time(tt <- weightedDetection(occWeights = visWeights,
                                    visWeights = visWeights,
                                    noOccur = 100,
                                    noLocations = seq(10, 100, by = 10),
                                    noVisits = 5
))


}

}
