% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOccProb.R
\name{createOccProb}
\alias{createOccProb}
\title{createOccProb}
\usage{
createOccProb(map, intercept = 0.5, formula = ~NULL, sigmaFylke = 0.1,
  sigmaKommune = 0.1, sigmaGrid = 0, nYears = 5, interceptTrend = -0.05,
  sortFylke = T, sortKommune = T, sortGrid = T)
}
\arguments{
\item{formula}{Not yet implemented}

\item{sigmaFylke}{Standard deviance of fylkes}

\item{sigmaFylke}{Standard deviance of kommunes.}

\item{sigmaFylke}{Standard deviance of grids.}
}
\description{
Creates a probability of species occurrence based on explanatory factors.
The levels are calculated additively with a fylke effect, a kommune effect, plus a random part (sigma) that
determines the intra-kommune variation among grid cells.
This is intended to fill a "map" with values of occurrences, and expects columns named fylke and kommune.
}
\examples{
\dontrun{

}
}
